modigied gmm
